requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.1
django-redis==5.4.0
djangorestframework==3.16.0
pillow==11.2.1
redis==6.1.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0